+++ R/RcppExports.R
@@ -93,8 +93,8 @@
     invisible(.Call(`_Cyclops_cyclopsSetParameterizedPrior`, inRcppCcdInterface, priorTypeName, priorFunction, startingParameters, excludeNumeric))
 }
 
-.cyclopsGetProfileLikelihood <- function(inRcppCcdInterface, inCovariate, points, threads, includePenalty) {
-    .Call(`_Cyclops_cyclopsGetProfileLikelihood`, inRcppCcdInterface, inCovariate, points, threads, includePenalty)
+.cyclopsGetProfileLikelihood <- function(inRcppCcdInterface, inCovariate, points, threads, includePenalty, returnDerivatives) {
+    .Call(`_Cyclops_cyclopsGetProfileLikelihood`, inRcppCcdInterface, inCovariate, points, threads, includePenalty, returnDerivatives)
 }
 
 .cyclopsProfileModel <- function(inRcppCcdInterface, sexpCovariates, threads, threshold, override, includePenalty) {
@@ -121,12 +121,24 @@
     .Call(`_Cyclops_cyclopsRunCrossValidationl`, inRcppCcdInterface)
 }
 
+.cyclopsCacheForJava <- function(inRcppCcdInterface) {
+    .Call(`_Cyclops_cyclopsCacheForJava`, inRcppCcdInterface)
+}
+
+.cyclopsClearCacheForJava <- function() {
+    invisible(.Call(`_Cyclops_cyclopsClearCacheForJava`))
+}
+
 .cyclopsFitModel <- function(inRcppCcdInterface) {
     .Call(`_Cyclops_cyclopsFitModel`, inRcppCcdInterface)
 }
 
 .cyclopsRunBootstrap <- function(inRcppCcdInterface, outFileName, treatmentId, replicates) {
     .Call(`_Cyclops_cyclopsRunBootstrap`, inRcppCcdInterface, outFileName, treatmentId, replicates)
+}
+
+.cyclopsGetLogLikelihoodGradient <- function(inRcppCcdInterface, index) {
+    .Call(`_Cyclops_cyclopsGetLogLikelihoodGradient`, inRcppCcdInterface, index)
 }
 
 .cyclopsLogModel <- function(inRcppCcdInterface) {